+++ main.py
@@ -1,5 +1,5 @@
 import pygame
-from src import engine, window
+from pynegin import engine, window
 from game import Game
 
 window = window.Window((800,600), "My game")
