+++ pynegin/components/text.py
