+++ pynegin/conf.py
@@ -1,0 +1,3 @@
+import os, sys
+
+HOME = os.path.dirname(os.path.realpath(sys.argv[0]))
