+++ pynegin/constants/colors.py
