+++ pynegin/container.py
