+++ pynegin/engine.py
@@ -1,4 +1,5 @@
 import pygame
+import os
 
 class Engine:
     def __init__(self, gameLogic, window, max_fps=30):
